+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 -   Run tests in GitHub
 -   Use prettier to format project
+-   Show all opening hours
 
 ### Fixed